Convert to SI units inside SMACCMPilot.Hardware drivers.
Note that the values coming from the hardware are signed, in
two's-complement representation, but were being processed in the
firmware as if they were unsigned. The Python script then re-interpreted
the telemetry bytestream as signed numbers, which worked fine until I
tried to do any kind of math on the firmware's representation.