disable flight and flight-hil by renaming their build.mk files.
I recently changed the way the ivory-bsp-stm32 code works (specifically, it generates a vector table and init code) so that it is no longer compatible with systems that depend on hwf4. flight and flight-hil are the last two apps that use both ivory-bsp-stm32 and hwf4. So, I’m going to disable them in the build in the development branch for now, because 1. we know both have serious issues and are not flight worthy on the current tower due to interrupt unschedulability, 2. we really don’t care about flight until tower is fixed and the blob is finally dead. hwf4 is only used by the blob.

The flight app has no business being built at this point, because it is known to be garbage. The blob is only half dead, the bsp is only half alive, and the failure of the flight app build is symptomatic of exactly that problem.

Put another way: anyone who hits make and gets a ‘flight’ binary out is will either consider that artifact meaningless (because both the bsp and blob are, in their own ways, broken) or be disappointed when they discover how broken it is. The first case adds no information, so we should optimize the second case.

I don’t want to patch up the build constantly to hide the details of the particular way the bsp and blob code fail to interact properly. Every other app in the tree is either blob only or bsp only, so flight is really just a weird special case at this point.

At some point the blob will be dead and tower changes will make the new bsp actually work. At that point we can revisit the C build of those apps, properly, because they won’t be stuck in limbo anymore.

If you want to use flight, you should use the -build master branch, Only. thats been the case for many months now, we’re just finally making it official.