Paul's Kalman filter was non-standard; generalize.
During the Kalman prediction step, Paul Riseborough's filter included
special handling to fold the current accelerometer and gyro measurements
into the process model. By augmenting the state vector with these
disturbances, and the prior covariance matrix with the disturbance
covariance, we can get the same result while using the general form of
the Kalman filter as given in textbooks or Wikipedia.

I suspect this is a standard technique, or at least Kalman filter
folklore, but I haven't been able to find a reference for it.

Note that I forgot to include process noise in previous commits. This
commit makes that explicit and gives me a place to add process noise
later.