ivory-px4-hw: rework each sensor driver to have generalized sample types
so instead of an mpu6000_sample it has an accelerometer_sample and
gyroscope_sample, and so on. Makes it easier to deal with multiple
platforms, e.g. the lsm303d vs the hmc5883l, having two accelerometers
on pixhawk, and optionally supporting the second gyro on pixhawk at a later
date.