Constructing a cleaner 'Galois Embedded Crypto'.
This is sort of an API from which we can construct the higher-level
"commsec like" functionallity.  There is currently a little bit of
abstraction muddying due to the gec_{encrypt,decrypt}() routine being
precisely the box() style abstraction (aka securePkg), but without
exposing a rather clumsy AEAD primitive, such as GCM, this is going
to just be the way it is.

I now need to update the commsec and commsec-ke code to use GEC, setup
the build system, and make some tests.