Revert to smaller tags and counter.
As discussed after commit c0f08219, we need to stick with 96 byte
messages on the wire.  I don't have any issue with the smaller tag size
given the flight lifetimes and parameters, so will stick with the
original design unless a need arises for larger messages, faster rates,
or longer flight times.

Recall the original plan is flights that last under 24 hours with
message rates in the ballpark of 100 msg/sec.  This means under 10M
messages for the lifetime of a key (which we could re-negotiate if it
comes to that), well under the 2^32 messages suggested for a 64 bit
GCM tag and the 32-bit counter limit.