smaccm-commsec: Change Message format.
The KE message format now no longer re-transmits duplicate public
ephemeral keys as ciphertext for key confirmation as the signature is
sufficient.  Not sure what I was thinking when I wrote it that way.

Also, there was a misuse of 'IV_LEN' for both CTR and GCM.  We now have
two *_IV_LEN macros for use in the different cases.