ivory-px4-hw: match px4 project spi settings
i can't be totally sure, but i believe the bug is now fixed

its funny because i tried these spi settings previously but they didn't "solve"
the problem. but perhaps, in combination with all the other changes i made,
they are now sufficient to solve the issue.

on the oscope, the spi bus still looks considerably worse than the same hardware
running ArduCopter. so, i'm not 100% confident we've seen the end of this bug.