Makefile.sandbox: Move backjumps out of SANDBOX_INSTALL.
Another chapter in the ongoing saga of the fight with stale versions of
cabal:

We've had issues where cabal-install would fail to find a satisfying
solution to the set of package dependencies, but only for older
versions. 1.18 was broken and 1.22 is fine, if I recall correctly.

On June 18th, Pat solved the problem by adding the --max-backjumps
option to our `cabal install` invocations, which makes older cabal
persistent enough in its search to find correct solutions. However I
wasn't satisfied with that and reverted it, after fixing some
dependencies to be easier to resolve.

On July 6th, Lee re-applied a similar patch--due to reports of lingering
issues from someone external who was stuck on older cabal, I think?--but
this time added the option to the SANDBOX_INSTALL makefile variable
instead of to the `cabal install` invocation itself.

Unfortunately, my Makefile.sandbox uses SANDBOX_INSTALL in a magic way
to support building smaccm-comm-schema, and this change caused the other
makefiles to not run `cabal build` any more. Somehow none of us noticed?

So with this commit, I'm keeping Lee's max-backjumps setting but
applying it the way Pat did it. Hopefully that will work for everyone.