Add a GIDL message and HUD button to reboot the Linux VM.
This isn't wired up in the Odroid yet, and the GCS HTTP server is
rejecting the button clicks with "400 Bad Request"--but from everything
I can check, this code should be correct. I'm committing it so Lee can
work on the Odroid side while I continue debugging the GCS side.