smaccm-flight: initial tuning of iris is identical to iris+
after fixing the pwm limits for the orig iris, it flies Much better
than with the iris+ pwm limits, which makes sense because all motor
outputs are scaled according to the limits.

I need to take this copter outside to get the tuning better than it is
now. It may be possible to improve some, but until it stops raining
and its hard to say.