improve LIDAR and PX4Flow drivers
Added timeout logic to these drivers, even though in theory they
shouldn't need it. For some reason, though, not every error condition
is making it back through to the driver, despite the scheduler getting
cleared of their pending flag. So now there is a generous 1sec timeout
after which they'll start trying to sample again.

If we hit asserts in the scheduler code, this is likely why, because
of the invariant of 1-1 requests<->responses, but I cannot see how
this situation is arising.