change arming law to require good sensor data
Unlike the other inputs to the arming law, the good sensor data is
only required when transitioning from safe to armed. This is so that
an occasional bad sensor reading while flying does not cause the
vehicle to be safed mid-air.

This commit also removes the last vestiges of the old calibration
scheme, as the previous arming law was the only thing still using
those values.